+++ app/src/main/java/com/intake/intakevisor/analyse/processor/model/Position.kt
@@ -3,4 +3,4 @@ package com.intake.intakevisor.analyse.processor.model
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-data class Position(val X: Int, val Y: Int)
+data class Position(val X: Double, val Y: Double)
